refactor(pane): Phase 3 — lift CreatePaneTask into pane/creation.rs
Moves the UI-thread browser creation task (`wrap_task! { pub struct
CreatePaneTask ... }`) from `browser_panes.rs` to a new
`agentmux-cef/src/pane/creation.rs`. Pure code motion, no behavior change.

Files
-----

`agentmux-cef/src/pane/creation.rs` (new):
- `CreatePaneTask` with identical `execute()` body.
- Same one-way dependencies as before:
    • `crate::state::AppState`
    • `crate::client::{AgentMuxHandler, AgentMuxClient}`
    • `crate::commands::window::find_own_top_level_window`
  None of these form cycles — `client::*` is only the pane handler type,
  and Phase 4 will finish splitting by lifting the pane callbacks out of
  `client.rs` altogether.

`agentmux-cef/src/pane/mod.rs`:
- `pub mod creation;` + `pub use creation::CreatePaneTask;`
- Trim the `PaneLifecycle` re-export — it's reachable via
  `crate::pane::lifecycle::PaneLifecycle` for the one in-module test
  that needs it. Keeps the warning list clean.
- Re-export `install_pane_focus_redirect` with `#[allow(unused_imports)]`
  and mark the function itself `#[allow(dead_code)]` — Phase 4 wires it
  in, don't want a warning shower in the meantime.

`agentmux-cef/src/browser_panes.rs`:
- `use crate::pane::{CreatePaneTask, ...};`
- ~80 lines of `wrap_task!` body removed.

Testing
-------

`cargo test` — 27/27 pass (same set as Phase 2). Build warnings: 18 now,
was 19 before the phase — net one fewer.

Next phase (the payoff)
-----------------------

Phase 4 lifts the pane-specific branches of `on_after_created`,
`on_before_close`, `on_set_focus`, and `on_load_end` out of `client.rs`
into `pane/callbacks.rs`. That's the phase that finally:
  - Breaks the `browser_panes ↔ client` import cycle.
  - Wires `install_pane_focus_redirect` into `on_after_created` +
    `on_load_end`, fixing the post-navigation focus-steal regression
    documented in `SPEC_BROWSER_PANE_LIFECYCLE.md` §5 race #5.
