+++ agentmux-cef/src/browser_panes.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 
 use cef::*;
 
-use crate::pane::{PaneStateMachine, RegisterResult};
+use crate::pane::{CreatePaneTask, PaneStateMachine, RegisterResult};
 use crate::state::AppState;
 
 /// Abstraction over the CEF-side operations `close()` performs on the host
@@ -320,89 +320,7 @@ impl BrowserPaneManager {
     }
 }
 
-// ── UI thread task: create browser ──────────────────────────────────────────
-
-wrap_task! {
-    pub struct CreatePaneTask {
-        state: Arc<AppState>,
-        block_id: String,
-        label: String,
-        url: String,
-        rect: Rect,
-    }
-
-    impl Task {
-        fn execute(&self) {
-            // Running on the CEF UI thread.
-
-            // Get parent HWND via Win32 enumeration (CEF Views returns null).
-            #[cfg(target_os = "windows")]
-            let parent_hwnd_raw = unsafe {
-                crate::commands::window::find_own_top_level_window()
-            };
-            #[cfg(not(target_os = "windows"))]
-            let parent_hwnd_raw: *mut std::ffi::c_void = std::ptr::null_mut();
-
-            if parent_hwnd_raw.is_null() {
-                tracing::error!(block_id = %self.block_id, "cannot find main window HWND — aborting browser pane creation");
-                return;
-            }
-
-            // Queue label for on_after_created registration (stored in state.browsers)
-            self.state.pending_window_labels.lock().push_back(self.label.clone());
-
-            let handler = crate::client::AgentMuxHandler::new_with_pane(self.state.clone(), 0, true);
-            let mut client = Some(crate::client::AgentMuxClient::new(handler, true));
-
-            let url_cef = CefString::from(self.url.as_str());
-            let settings = BrowserSettings::default();
-
-            #[cfg(target_os = "windows")]
-            let window_info = {
-                let parent_hwnd = sys::HWND(parent_hwnd_raw as *mut _);
-                // Use the clean set_as_child helper — it fills style/parent/bounds
-                // correctly and leaves other fields zeroed (in particular `window`
-                // which is an OUTPUT field filled by CEF).
-                let mut wi = WindowInfo::default().set_as_child(parent_hwnd, &self.rect);
-                // Match the main process runtime style (ALLOY throughout the app).
-                wi.runtime_style = RuntimeStyle::ALLOY;
-                wi
-            };
-
-            #[cfg(not(target_os = "windows"))]
-            {
-                tracing::warn!(block_id = %self.block_id, "browser panes not yet implemented on this platform");
-                return;
-            }
-
-            #[cfg(target_os = "windows")]
-            {
-                let result = browser_host_create_browser(
-                    Some(&window_info),
-                    client.as_mut(),
-                    Some(&url_cef),
-                    Some(&settings),
-                    None, // extra_info
-                    None, // request_context
-                );
-
-                if result == 0 {
-                    tracing::error!(block_id = %self.block_id, "browser_host_create_browser returned 0");
-                    return;
-                }
-
-                tracing::info!(
-                    block_id = %self.block_id,
-                    label = %self.label,
-                    url = %self.url,
-                    x = self.rect.x, y = self.rect.y,
-                    w = self.rect.width, h = self.rect.height,
-                    "browser pane created on UI thread"
-                );
-            }
-        }
-    }
-}
+// `CreatePaneTask` moved to `crate::pane::creation` in Phase 3.
 
 // ── Tests ───────────────────────────────────────────────────────────────────
 //
