+++ agentmux-cef/src/pane/mod.rs
@@ -10,11 +10,25 @@
 //!
 //! Re-exports the public surface that `browser_panes.rs` and tests consume.
 
+pub mod creation;
 pub mod lifecycle;
 #[cfg(target_os = "windows")]
 pub mod hwnd;
 
-pub use lifecycle::{PaneLifecycle, PaneStateMachine, RegisterResult};
+pub use creation::CreatePaneTask;
+pub use lifecycle::{PaneStateMachine, RegisterResult};
+
+// PaneLifecycle is reachable via `crate::pane::lifecycle::PaneLifecycle`
+// for tests that need to assert state; no re-export needed.
+#[cfg(test)]
+pub use lifecycle::PaneLifecycle;
 
 #[cfg(target_os = "windows")]
-pub use hwnd::{ALLOW_PANE_FOCUS_ONCE, install_pane_focus_redirect};
+pub use hwnd::ALLOW_PANE_FOCUS_ONCE;
+
+// install_pane_focus_redirect is defined but not yet wired to any caller;
+// Phase 4 will invoke it from pane `on_after_created` / `on_load_end`.
+// Re-export is ready; flagging suppressed until that lands.
+#[cfg(target_os = "windows")]
+#[allow(unused_imports)]
+pub use hwnd::install_pane_focus_redirect;
